requirements.txt
absl-py==0.8.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
google-auth==1.7.1
google-auth-oauthlib==0.4.1
grpcio==1.25.0
idna==2.8
Markdown==3.1.1
numpy==1.17.4
oauthlib==3.1.0
Pillow==6.2.1
protobuf==3.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
tensorboard==2.0.2
torch==1.3.1
torchvision==0.4.2
tqdm==4.39.0
urllib3==1.25.7
Werkzeug==0.16.0
